LPS-166968 Simplify, inline the method getDataLayoutConfigJSONObject as it is only used by one place
