+++ modules/apps/data-engine/data-engine-taglib/src/main/java/com/liferay/data/engine/taglib/internal/servlet/taglib/util/DataLayoutTaglibUtil.java
@@ -45,7 +45,6 @@ import com.liferay.dynamic.data.mapping.model.UnlocalizedValue;
 import com.liferay.dynamic.data.mapping.model.Value;
 import com.liferay.dynamic.data.mapping.service.DDMStructureLayoutLocalService;
 import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
-import com.liferay.dynamic.data.mapping.spi.form.builder.settings.DDMFormBuilderSettingsRetrieverHelper;
 import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
 import com.liferay.dynamic.data.mapping.storage.DDMFormValues;
 import com.liferay.dynamic.data.mapping.util.DDMFormFactory;
@@ -53,7 +52,6 @@ import com.liferay.dynamic.data.mapping.util.DDMFormLayoutFactory;
 import com.liferay.petra.function.UnsafeConsumer;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.json.JSONArray;
-import com.liferay.portal.kernel.json.JSONException;
 import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.json.JSONUtil;
@@ -114,64 +112,6 @@ public class DataLayoutTaglibUtil {
 
 		return _dataLayoutTaglibUtil._getDataDefinition(
 			dataDefinitionId, httpServletRequest);
-	}
-
-	public static JSONObject getDataLayoutConfigJSONObject(
-		String contentType, Locale locale) {
-
-		DataLayoutBuilderDefinition dataLayoutBuilderDefinition =
-			_getDataLayoutBuilderDefinition(contentType);
-
-		JSONObject dataLayoutConfigJSONObject = JSONUtil.put(
-			"allowFieldSets", dataLayoutBuilderDefinition.allowFieldSets()
-		).put(
-			"allowMultiplePages",
-			dataLayoutBuilderDefinition.allowMultiplePages()
-		).put(
-			"allowNestedFields", dataLayoutBuilderDefinition.allowNestedFields()
-		).put(
-			"allowRules", dataLayoutBuilderDefinition.allowRules()
-		).put(
-			"allowSuccessPage", dataLayoutBuilderDefinition.allowSuccessPage()
-		).put(
-			"disabledProperties",
-			dataLayoutBuilderDefinition.getDisabledProperties()
-		).put(
-			"disabledTabs", dataLayoutBuilderDefinition.getDisabledTabs()
-		).put(
-			"visibleProperties",
-			dataLayoutBuilderDefinition.getVisibleProperties()
-		);
-
-		if (dataLayoutBuilderDefinition.allowRules()) {
-			try {
-				dataLayoutConfigJSONObject.put(
-					"ruleSettings",
-					JSONUtil.put(
-						"dataProviderInstanceParameterSettingsURL",
-						_dataLayoutTaglibUtil.
-							_getDDMDataProviderInstanceParameterSettingsURL()
-					).put(
-						"dataProviderInstancesURL",
-						_dataLayoutTaglibUtil._getDDMDataProviderInstancesURL()
-					).put(
-						"functionsMetadata",
-						_dataLayoutTaglibUtil._getFunctionsMetadataJSONObject(
-							locale)
-					).put(
-						"functionsURL", _dataLayoutTaglibUtil._getFunctionsURL()
-					));
-			}
-			catch (JSONException jsonException) {
-				_log.error(jsonException);
-			}
-		}
-
-		dataLayoutConfigJSONObject.put(
-			"unimplementedProperties",
-			dataLayoutBuilderDefinition.getUnimplementedProperties());
-
-		return dataLayoutConfigJSONObject;
 	}
 
 	public static JSONObject getDataLayoutJSONObject(
@@ -243,19 +183,6 @@ public class DataLayoutTaglibUtil {
 		}
 
 		_dataLayoutBuilderDefinitions.remove(contentType);
-	}
-
-	private static DataLayoutBuilderDefinition _getDataLayoutBuilderDefinition(
-		String contentType) {
-
-		DataLayoutBuilderDefinition dataLayoutBuilderDefinition =
-			_dataLayoutBuilderDefinitions.get(contentType);
-
-		if (dataLayoutBuilderDefinition == null) {
-			return _dataLayoutBuilderDefinitions.get("default");
-		}
-
-		return dataLayoutBuilderDefinition;
 	}
 
 	private Set<Locale> _getAvailableLocales(
@@ -334,6 +261,19 @@ public class DataLayoutTaglibUtil {
 		return dataLayoutResource.getDataLayout(dataLayoutId);
 	}
 
+	private DataLayoutBuilderDefinition _getDataLayoutBuilderDefinition(
+		String contentType) {
+
+		DataLayoutBuilderDefinition dataLayoutBuilderDefinition =
+			_dataLayoutBuilderDefinitions.get(contentType);
+
+		if (dataLayoutBuilderDefinition == null) {
+			return _dataLayoutBuilderDefinitions.get("default");
+		}
+
+		return dataLayoutBuilderDefinition;
+	}
+
 	private JSONObject _getDataLayoutJSONObject(
 		Set<Locale> availableLocales, String contentType, Long dataDefinitionId,
 		Long dataLayoutId, HttpServletRequest httpServletRequest,
@@ -379,16 +319,6 @@ public class DataLayoutTaglibUtil {
 
 			return _jsonFactory.createJSONObject();
 		}
-	}
-
-	private String _getDDMDataProviderInstanceParameterSettingsURL() {
-		return _ddmFormBuilderSettingsRetrieverHelper.
-			getDDMDataProviderInstanceParameterSettingsURL();
-	}
-
-	private String _getDDMDataProviderInstancesURL() {
-		return _ddmFormBuilderSettingsRetrieverHelper.
-			getDDMDataProviderInstancesURL();
 	}
 
 	private Long _getDefaultDataLayoutId(
@@ -463,18 +393,6 @@ public class DataLayoutTaglibUtil {
 		}
 	}
 
-	private JSONObject _getFunctionsMetadataJSONObject(Locale locale)
-		throws JSONException {
-
-		return _jsonFactory.createJSONObject(
-			_ddmFormBuilderSettingsRetrieverHelper.
-				getSerializedDDMExpressionFunctionsMetadata(locale));
-	}
-
-	private String _getFunctionsURL() {
-		return _ddmFormBuilderSettingsRetrieverHelper.getDDMFunctionsURL();
-	}
-
 	private void _setFieldIndexTypeNone(JSONObject jsonObject) {
 		for (JSONObject pageJSONObject :
 				(Iterable<JSONObject>)jsonObject.getJSONArray("pages")) {
@@ -516,10 +434,6 @@ public class DataLayoutTaglibUtil {
 
 	@Reference
 	private DataLayoutResource.Factory _dataLayoutResourceFactory;
-
-	@Reference
-	private DDMFormBuilderSettingsRetrieverHelper
-		_ddmFormBuilderSettingsRetrieverHelper;
 
 	@Reference
 	private DDMFormFieldTypeServicesRegistry _ddmFormFieldTypeServicesRegistry;
